Update duration measure
Although we cannot directly observe video duration, we can use the length of the video as  a proxy to get a rough estimate of time spent watching if students watched the entire videos (likely an overestimate).

This commit also contains a fair bit of content added to the Study Design section, especially the empirical strategies, and a proof of standard error calculations for the Neyman repeated sampling approach.