Results, attrition, balance, instrument relevance, abstract, intro rewrite (#6)
* Update duration measure

Although we cannot directly observe video duration, we can use the length of the video as  a proxy to get a rough estimate of time spent watching if students watched the entire videos (likely an overestimate).

This commit also contains a fair bit of content added to the Study Design section, especially the empirical strategies, and a proof of standard error calculations for the Neyman repeated sampling approach.

* Attrition, balance, instrument relevance, abstract, intro redux

Lots of text added to the paper. Some small stuff in tables and plots, including:
- added mid2 to the cdf
- rename PSET 'tutoring" and Piazza "discussion board"

Added an attrition flowchart that makes it easy to see where students attrited.

Removed the incorrect duration description and simplified the LATE empirical strategy to be more in line with existing literature.

Next To-do's:
1. Add text to results and discussion
2. Move theory section to after the results
3. Add Fryer 2011 to lit review
4. Add some blurb about control students watching a lot of videos and the ATE being greater than the LATE

* Results section

First draft of results section. Calculated Anderson-Rubin confidence intervals. Adjusted studying spillover table for clarity.

Thinking about adding a benchmarking section, perhaps as an alternative to the lit review section (or at least shrink it). Still need to move theory section.