Julie comments, remove RD section, move heterogeneity section
Because we can examine ITTs along the midterm 1 distribution below the median (including at the median), it doesn't make sense to spend a large chunk of the paper on an RD description and analysis. As such I'm opting to drop the RD section in favor of a modestly longer heterogeneity section.

I also misplaced a reference the heterogeneity results, so I move that to the appropriate location. I am also moving the heterogeneity table results to the main body of the paper (instead of appendix) for now.

I also make some stylistic changes and condense content where appropriate.